Rename library package and clean up
