+++ ComposeTesting/sample/build.gradle
@@ -54,25 +54,12 @@ android {
 }
 
 dependencies {
-    implementation project(path: ':library')
-
-    implementation kotlinDependencies.kotlinStdlib
-    implementation androidxDependencies.ktxCore
-    implementation androidxDependencies.appCompat
-
-    implementation composeDependencies.composeUI
     implementation composeDependencies.composeMaterial
-    implementation composeDependencies.composeUITooling
-    implementation composeDependencies.activityCompose
 
     implementation googleDependencies.material
+
     implementation microsoftDependencies.twoPaneLayout
 
-    androidTestImplementation(testDependencies.androidxTestCore)
-    androidTestImplementation(testDependencies.androidxTestRules)
-    androidTestImplementation(testDependencies.androidxTestRunner)
-    androidTestImplementation(testDependencies.composeUITest)
+    androidTestImplementation project(path: ':library')
     androidTestImplementation(testDependencies.composeJunit)
-    androidTestImplementation(testDependencies.uiAutomator)
-    debugImplementation(testDependencies.composeUITestManifest)
 }
