+++ ComposeTesting/sample/src/androidTest/java/com/microsoft/device/dualscreen/testing/sample/FoldingFeatureTest.kt
@@ -11,17 +11,16 @@ import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import com.microsoft.device.dualscreen.testing.createWindowLayoutInfoPublisherRule
-import com.microsoft.device.dualscreen.testing.sample.R
-import com.microsoft.device.dualscreen.testing.getString
-import com.microsoft.device.dualscreen.testing.simulateHorizontalFoldingFeature
-import com.microsoft.device.dualscreen.testing.simulateVerticalFoldingFeature
+import com.microsoft.device.dualscreen.testing.compose.getString
 import com.microsoft.device.dualscreen.testing.sample.ui.theme.ComposeTestingTheme
+import com.microsoft.device.dualscreen.testing.compose.simulateHorizontalFoldingFeature
+import com.microsoft.device.dualscreen.testing.compose.simulateVerticalFoldingFeature
 import org.junit.Rule
 import org.junit.Test
 import org.junit.rules.RuleChain
 import org.junit.rules.TestRule
 
-class FoldingFeatureHelperTest {
+class FoldingFeatureTest {
     private val composeTestRule = createAndroidComposeRule<MainActivity>()
     private val publisherRule = createWindowLayoutInfoPublisherRule()
